# optimize layernorm forward